requirements.txt
# AI-Native Skills Architecture Requirements
# Complete dependency list for all 10 skills

# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM & Embeddings
openai==1.3.7
cohere==4.37
python-dotenv==1.0.0

# Vector Database
qdrant-client==2.7.0

# Document Processing & RAG
python-multipart==0.0.6

# Database
psycopg[binary]==3.1.15
sqlalchemy==2.0.23
asyncpg==0.29.0

# Logging
python-json-logger==2.0.7

# Git Operations (system git required)
GitPython==3.1.40

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Data Processing
numpy==1.24.3
pandas==2.1.3

# System & Utilities
requests==2.31.0
click==8.1.7
colorama==0.4.6
